Use iterative JavaScript regex fallback scanner
