+++ AdvancedCore/src/main/java/com/bencodez/advancedcore/api/javascript/JavascriptPlaceholderBinder.java
@@ -46,7 +46,6 @@ public final class JavascriptPlaceholderBinder {
     private static final String TREE_PACKAGE = "org.openjdk.nashorn.api.tree";
     private static final Pattern FALLBACK_STRING = Pattern.compile("'(?:\\\\.|[^'\\\\])*'|\"(?:\\\\.|[^\"\\\\])*\"");
     private static final Pattern FALLBACK_TEMPLATE = Pattern.compile("`(?:\\\\.|[^`\\\\])*`");
-    private static final Pattern FALLBACK_REGEX = Pattern.compile("/(?:\\\\.|[^/\\\\\\r\\n])+/[dgimsuvy]*");
 
     private JavascriptPlaceholderBinder() {
     }
@@ -353,9 +352,57 @@ public final class JavascriptPlaceholderBinder {
                 addFallbackTemplateExpressions(source, template, contexts.templateExpressions);
             }
 
-            addPatternRanges(source, FALLBACK_REGEX, contexts.regexes, contexts);
+            addFallbackRegexRanges(source, contexts);
             contexts.sort();
             return contexts;
+        }
+
+        private static void addFallbackRegexRanges(String source, JavascriptContexts contexts) {
+            for (int i = 0; i < source.length(); i++) {
+                if (source.charAt(i) != '/' || contexts.containing(contexts.strings, i) != null
+                        || contexts.containing(contexts.templates, i) != null) {
+                    continue;
+                }
+
+                boolean escaped = false;
+                boolean characterClass = false;
+                for (int j = i + 1; j < source.length(); j++) {
+                    char current = source.charAt(j);
+                    if (current == '\r' || current == '\n') {
+                        break;
+                    }
+                    if (escaped) {
+                        escaped = false;
+                        continue;
+                    }
+                    if (current == '\\') {
+                        escaped = true;
+                        continue;
+                    }
+                    if (current == '[') {
+                        characterClass = true;
+                        continue;
+                    }
+                    if (current == ']') {
+                        characterClass = false;
+                        continue;
+                    }
+                    if (current != '/' || characterClass) {
+                        continue;
+                    }
+
+                    int end = j + 1;
+                    while (end < source.length() && "dgimsuvy".indexOf(source.charAt(end)) >= 0) {
+                        end++;
+                    }
+                    Range candidate = new Range(i, end);
+                    if (!contexts.overlapsLiteral(candidate)) {
+                        contexts.regexes.add(candidate);
+                    }
+                    i = end - 1;
+                    break;
+                }
+            }
         }
 
         private static void addPatternRanges(String source, Pattern pattern, List<Range> target,
@@ -439,8 +486,8 @@ public final class JavascriptPlaceholderBinder {
                 return downloaded;
             }
 
-            // nashorn-core is packaged with AdvancedCore so parser support remains
-            // available even when the active ScriptEngine is Rhino/GraalJS.
+            // Use a parser already visible to AdvancedCore when available. If not,
+            // the handler can prepare a dedicated runtime Nashorn parser loader below.
             ClassLoader own = JavascriptPlaceholderBinder.class.getClassLoader();
             if (canLoadParser(own)) {
                 return own;
